fix: Small cleanups to the benefits page.
I removed the "min 10 PTO per year" which was previously included
as a value statement (we want you to take care of yourself), but
could be read as something that I'm counting and enforcing
(which I'm not).

Otherwise I mostly cleaned up some of the (previously copied)
language in the expenses section from GitLab, as well as fixed
one typo and made explicit our 11 holidays we observed last year.